- Change btcHeightWhenPegoutTxIndexActivates type to int type. - Update test to assert getBtcHeightWhenPegoutTxIndexActivates returns the expected value for each network. - set btcHeightWhenPegoutTxIndexActivates mock values for each network.

+++ rskj-core/src/main/java/co/rsk/config/BridgeConstants.java
@@ -86,7 +86,7 @@ public abstract class BridgeConstants {
 
     protected int numberOfBlocksBetweenPegouts;
 
-    protected long btcHeightWhenPegoutTxIndexActivates;
+    protected int btcHeightWhenPegoutTxIndexActivates;
 
     public NetworkParameters getBtcParams() {
         return NetworkParameters.fromID(btcParamsString);
@@ -188,7 +188,7 @@ public abstract class BridgeConstants {
         return numberOfBlocksBetweenPegouts;
     }
 
-    public long getBtcHeightWhenPegoutTxIndexActivates() {
+    public int getBtcHeightWhenPegoutTxIndexActivates() {
         return btcHeightWhenPegoutTxIndexActivates;
     }
 }
